+++ btclib/taproot.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python3
+# !/usr/bin/env python3
 
 # Copyright (C) 2021 The btclib developers
 #
@@ -8,29 +8,31 @@
 # No part of btclib including this file, may be copied, modified, propagated,
 # or distributed except according to the terms contained in the LICENSE file.
 
-from .curve import Curve
-from .curvemult import mult
-from .curves import secp256k1
-from .hashes import tagged_hash
-from .sighash import SegwitV1SignatureHash
-from .ssa import _sign
+
+"""Taproot related functions"""
+
+
+from btclib.ecc.curve import Curve, mult, secp256k1
+from btclib.ecc.ssa import sign_
+from btclib.hashes import tagged_hash
+from btclib.sig_hash import taproot
 
 
 def tweak_pubkey(pubkey, h, ec: Curve = secp256k1):
-    t = int.from_bytes(tagged_hash("TapTweak", pubkey + h), "big")
+    t = int.from_bytes(tagged_hash(b"TapTweak", pubkey + h), "big")
     if t >= ec.n:
         raise ValueError
     x = int.from_bytes(pubkey, "big")
-    Q = ec.add((x, ec.y_odd(x, 0)), mult(t))
-    has_even_y = ec.y_odd(Q[0]) != Q[1]
+    Q = ec.add((x, ec.y_even(x)), mult(t))
+    has_even_y = ec.y_even(Q[0]) != Q[1]
     return 0 if has_even_y else 1, Q[0].to_bytes(32, "big")
 
 
 def taproot_tweak_seckey(seckey0, h, ec: Curve = secp256k1):
     P = mult(int.from_bytes(seckey0, "big"))
-    has_even_y = ec.y_odd(P[0]) != P[1]
+    has_even_y = ec.y_even(P[0]) != P[1]
     seckey = seckey0 if has_even_y else ec.n - seckey0
-    t = int.from_bytes(tagged_hash("TapTweak", P[0].to_bytes(32, "big") + h), "big")
+    t = int.from_bytes(tagged_hash(b"TapTweak", P[0].to_bytes(32, "big") + h), "big")
     if t >= ec.n:
         raise ValueError
     return (seckey + t) % ec.n
@@ -39,14 +41,14 @@ def taproot_tweak_seckey(seckey0, h, ec: Curve = secp256k1):
 def taproot_tree_helper(script_tree):
     if isinstance(script_tree, tuple):
         leaf_version, script = script_tree
-        h = tagged_hash("TapLeaf", bytes([leaf_version]) + script.serialize())
+        h = tagged_hash(b"TapLeaf", bytes([leaf_version]) + script.serialize())
         return ([((leaf_version, script), bytes())], h)
     left, left_h = taproot_tree_helper(script_tree[0])
     right, right_h = taproot_tree_helper(script_tree[1])
     ret = [(l, c + right_h) for l, c in left] + [(l, c + left_h) for l, c in right]
     if right_h < left_h:
         left_h, right_h = right_h, left_h
-    return (ret, tagged_hash("TapBranch", left_h + right_h))
+    return (ret, tagged_hash(b"TapBranch", left_h + right_h))
 
 
 def taproot_output_script(internal_pubkey, script_tree):
@@ -67,7 +69,7 @@ def taproot_output_script(internal_pubkey, script_tree):
 def taproot_sign_key(script_tree, internal_seckey, hash_type):
     _, h = taproot_tree_helper(script_tree)
     output_seckey = taproot_tweak_seckey(internal_seckey, h)
-    sig = _sign(SegwitV1SignatureHash(hash_type), output_seckey)
+    sig = sign_(taproot(hash_type), output_seckey)
     if hash_type != 0:
         sig += bytes([hash_type])
     return [sig]
@@ -78,4 +80,5 @@ def taproot_sign_script(internal_pubkey, script_tree, script_num, inputs):
     (leaf_version, script), path = info[script_num]
     output_pubkey_y_parity, _ = tweak_pubkey(internal_pubkey, h)
     pubkey_data = bytes([output_pubkey_y_parity + leaf_version]) + internal_pubkey
+    return inputs + [script, pubkey_data + path]
     return inputs + [script, pubkey_data + path]
