+++ btclib/tx/block_header.py
@@ -21,7 +21,8 @@ from typing import Any, Dict, List, Mapping, Tuple, Type, TypeVar, Union
 
 from btclib.alias import BinaryData, Octets
 from btclib.exceptions import BTClibValueError
-from btclib.utils import bytes_from_octets, bytesio_from_binarydata, hash256
+from btclib.hashes import hash256
+from btclib.utils import bytes_from_octets, bytesio_from_binarydata
 
 # python 3.6
 if sys.version_info.minor == 6:  # pragma: no cover
