+++ tests/psbt/test_psbt.py
@@ -17,6 +17,7 @@ import pytest
 
 from btclib.ecc import der, dsa, sec_point
 from btclib.exceptions import BTClibValueError
+from btclib.hashes import hash160, hash256, ripemd160, sha256
 from btclib.psbt.psbt import (
     PSBT_DELIMITER,
     PSBT_SEPARATOR,
@@ -30,7 +31,6 @@ from btclib.script.witness import Witness
 from btclib.tx.tx import Tx
 from btclib.tx.tx_in import OutPoint, TxIn
 from btclib.tx.tx_out import TxOut
-from btclib.utils import hash160, hash256, ripemd160, sha256
 
 # first tests are part of the official BIP174 test vectors
 
@@ -319,7 +319,9 @@ def test_valid_sign_2() -> None:
     )
     assert psbt.inputs[0].witness_utxo is not None
     # pylance cannot grok the following line, even considering the above line
-    transaction = Tx(1, 2, [transaction_input], [psbt.inputs[0].witness_utxo])  # type: ignore
+    transaction = Tx(
+        1, 2, [transaction_input], [psbt.inputs[0].witness_utxo]
+    )  # type: ignore
     psbt.tx.vin[0].prev_out.tx_id = transaction.id
     psbt.inputs[0].non_witness_utxo = transaction
     psbt.assert_signable()
